codecrafters submit [skip ci]

+++ src/ItemValue.cs
@@ -6,16 +6,16 @@ public record ItemValue
     private double createdAt;
     private bool hasExpiration; 
 
-    public ItemValue(string value, int timeToLive = int.MaxValue)
+    public ItemValue(string value, double timeToLive = double.MaxValue)
     {
         Value = value;
         TimeToLive = timeToLive;
         createdAt = GetCurrentMiliseconds();
-        hasExpiration = (timeToLive != int.MaxValue);
+        hasExpiration = (timeToLive != double.MaxValue);
     }
 
     public string Value { get; set; }
     public double ValidUntil { get => (hasExpiration) ? createdAt + TimeToLive : double.MaxValue; }
-    public int TimeToLive { get; set; }
+    public double TimeToLive { get; set; }
     public bool IsValid => ValidUntil > GetCurrentMiliseconds();
 }
